backend/requirements.txt
Django==4.2.1
django-bootstrap-v5==1.0.11
django-cors-headers==4.1.0
djangorestframework==3.14.0